Add client-side routing for the three data views (#46)
The router that reads the value of the view query parameter from the url, and finds a corresponding view
from filters config.

If the corresponding view was not found (i.e. the url does not have a view parameter, or the value of the view
parameter does not match any of the views in the config), then the first view from the config is used,
and the url is updated to match that view.

When views are changed, the url is updated accordingly. The updates to the url are done through the history.replace
method, such that pressing the browser back button will take the user to the previous page, rather than the previous view.